refactor: move tool schemas to core/tool_schemas.py
Extract TOOL_SCHEMAS, NL_TOOL_SCHEMAS, NL_SCHEMA_MAP, TOOL_CATEGORIES,
and get_tool_schemas() from tools.py into a dedicated core module. This
fixes the layer violation where turing_protocol.py (core) imported private
names from tools.py (top-level).

The previously private _NL_SCHEMA_MAP is now public as NL_SCHEMA_MAP.
tools.py re-exports all names for backwards compatibility.

Closes #87
